requirements.txt
# 웹 프레임워크
Flask==3.0.0

# 데이터 처리
numpy==1.24.3
pandas==2.0.3

# 머신러닝
scikit-learn==1.3.2
joblib==1.3.2

# 컴퓨터 비전
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# 시각화
matplotlib==3.8.2

# XAI (Explainable AI)
shap==0.43.0
lime==0.2.0.1

# 딥러닝 (PyTorch는 별도 설치 필요)
# GPU 버전 (CUDA 12.1): pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# CPU 버전: pip install torch torchvision torchaudio
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
tqdm>=4.65.0

